どんな色で描画するかはcameraの仕事
